Phase 1 bit-exact — float64 intermediates; r_parity floors tightened to 1e-10
8 Phase 1 steps now max_diff < 1e-10 vs R infercnv on oligodendroglioma:
  step03 normalize    (was 1e-2, now 4.9e-11)
  step04 log2_plus1   (was 1e-5, now 5.3e-14)
  step08 subtract_ref (was 1e-3, now bit-exact)
  step09 max_threshold(was 1e-6, now 0)
  step11 center_cells (was 1e-4, now bit-exact)
  step14 invert_log2  (was 1e-4, now 5.6e-15)
  step16 prune_outliers(was 1e-4, now 4.2e-15)
  step02 filter_genes (was and remains bit-exact)

step10 smooth_pyramidinal stays at 1e-3 (scipy uniform_filter1d
accumulation order vs R); documented in NAMESPACE_PARITY as
"4 approximate"; hand-rolled R-matching convolution deferred.

Public contract preserved:
  result.cnv_matrix    float32 (cast at pipeline.py result assembly)
  result.cnv_matrix_fc float32
  All Phase 1 leaf functions now return float64.

Three unit-test dtype assertions rewritten to use
np.issubdtype(x, np.floating) — the float32 output pin was a legacy
contract, now enforced at result.cnv_matrix only.

HMM i3 end-to-end Jaccard on oligo UNCHANGED at 0.976 — Phase 2 still
consumes float32 via result.cnv_matrix. Plumbing float64 through to
Phase 2 is the next-session work; full plan in
docs/superpowers/HANDOFF_bit_exact.md §3 (recommended: add optional
result.cnv_matrix_f64 companion field; minimal public-contract disruption).

209 tests pass. Wheel py3-none-any rebuilt, twine PASSED.
No version bump, no tag, no push, no upload.

+++ pyinfercnv/cna/outlier_prune.py
@@ -42,10 +42,10 @@ def prune_outliers(
     if lower_bound is not None or upper_bound is not None:
         lo = -np.inf if lower_bound is None else float(lower_bound)
         hi = np.inf if upper_bound is None else float(upper_bound)
-        return np.clip(X, lo, hi).astype(np.float32)
+        return np.clip(X, lo, hi).astype(np.float64)
     if method == "average_bound":
         lo, hi = _average_bounds(X)
-        return np.clip(X, lo, hi).astype(np.float32)
+        return np.clip(X, lo, hi).astype(np.float64)
     if method is None:
         return np.asarray(X, dtype=np.float32).copy()
     raise ValueError(f"unknown outlier method {method!r}")
